Working 00, 01, 01b, 02
