+++ S_01_train_and_tunner_.py
@@ -10,7 +10,7 @@ import glob
 from technical_parameters_konk_tools import *
 from technical_parameters_konk_tools_Prepro import *
 
-print(dtreeviz.__version__)
+print("dtreevis version: ",dtreeviz.__version__)
 from matplotlib import pyplot as plt
 import matplotlib
 matplotlib.use('TKAgg')
